+++ src/main/java/com/demo/waterSupply/repository/CityRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CityRepository extends JpaRepository<CityModel, Integer> {
+public interface CityRepository extends JpaRepository<CityModel, Long> {
 
     CityModel findByCityName(String cityName);
 }
