+++ src/main/java/com/demo/waterSupply/repository/MeterReadingRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MeterReadingRepository extends JpaRepository<MeterReading, Integer> {
+public interface MeterReadingRepository extends JpaRepository<MeterReading, Long> {
 
 }
