+++ src/main/java/com/demo/waterSupply/repository/PermissionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PermissionRepository extends JpaRepository<PermissionModel, Integer> {
+public interface PermissionRepository extends JpaRepository<PermissionModel, Long> {
 }
