+++ src/main/java/com/demo/waterSupply/repository/RoleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends JpaRepository<RoleModel, Integer> {
+public interface RoleRepository extends JpaRepository<RoleModel, Long> {
 
 
     RoleModel findByRoleName(String roleName);
